move propertyKey "implementations" to own class
